2022.07.01 jquery, views, bootstrap updates

+++ app/admin_views.py
@@ -1,6 +1,6 @@
 from flask_sqlalchemy import SQLAlchemy
 from app import app
-from flask import Blueprint, flash, jsonify, redirect, render_template, send_file, Flask, send_from_directory, Response, url_for, request
+from flask import Blueprint, flash, jsonify, make_response, redirect, render_template, send_file, Flask, send_from_directory, Response, url_for, request
 from app.models import InvoiceModel, ClientModel, db
 from sqlalchemy import Boolean, create_engine, insert, select, func, distinct, true, values, delete
 from sqlalchemy.orm import sessionmaker, declarative_base
@@ -9,6 +9,7 @@ from matplotlib.pyplot import savefig
 from matplotlib.backends.backend_agg import FigureCanvasAgg as FigureCanvas
 from matplotlib.figure import Figure
 import io
+import json
 
 #bp = Blueprint('invoices', __name__, url_prefix='/dashboard')
 
@@ -32,10 +33,34 @@ collected_total_query = session.query(
 collected_total = round(collected_total_query.scalar(), 2)
 
 
+
+
+
 @app.route('/dashboard')
 def dashboard_index():
     return render_template('admin/templates/dashboard_template.html', paid_inv=paid_inv, unpaid_inv=unpaid_inv, total_inv=total_inv,
                            total_clients=total_clients, billed_total=billed_total, collected_total=collected_total)
+
+
+
+
+
+
+@app.route('/dashboard/all', methods=['GET', 'POST'])
+def invoice_index():
+    invoices = InvoiceModel.query.all()
+    invoice_list = []
+    for i in invoices:
+        invoice_list.append(i.serialize())
+    client_select = ClientModel.query.all()
+    client_list = []
+    for c in client_select:
+        client_list.append(c.serialize())
+    return render_template('admin/all.html', invoice_list=invoice_list, client_list=client_list, paid_inv=paid_inv, unpaid_inv=unpaid_inv, total_inv=total_inv,
+                           total_clients=total_clients, billed_total=billed_total, collected_total=collected_total)
+
+
+
 
 
 @app.route('/dashboard/unpaid', methods=['GET', 'POST'])
@@ -44,7 +69,7 @@ def unpaid_view():
     unpaid_list = []
     for u in unpaid:
         unpaid_list.append(u.serialize())
-    client_select = InvoiceModel.query.all()
+    client_select = ClientModel.query.all()
     client_list = []
     for c in client_select:
         client_list.append(c.serialize())
@@ -52,36 +77,50 @@ def unpaid_view():
                            total_clients=total_clients, billed_total=billed_total, collected_total=collected_total)
 
 
-@app.route('/dashboard/update/<int:inv_id>', methods=['GET', 'POST'])
-def update_invoice(inv_id: int):
+
+
+
+@app.route('/dashboard/inv-prepop/<int:inv_id>', methods=['GET', 'POST'])
+def inv_passthrough(inv_id: int):
     inv = InvoiceModel.query.filter(InvoiceModel.id == inv_id)
     inv_to_update = []
     for i in inv:
         inv_to_update.append(i.serialize())
+    invoice = inv_to_update[0]['id']
+    client_id = inv_to_update[0]['client']
 
     client_select = ClientModel.query.all()
     client_list = []
     for c in client_select:
         client_list.append(c.serialize())
-    print(inv_to_update)
+
     unpaid = InvoiceModel.query.filter(InvoiceModel.paid == False)
     unpaid_list = []
     for u in unpaid:
         unpaid_list.append(u.serialize())
-    return render_template('admin/update_inv.html', inv_to_update=inv_to_update, unpaid_list=unpaid_list, client_list=client_list, total_clients=total_clients, paid_inv=paid_inv, unpaid_inv=unpaid_inv, total_inv=total_inv,
+
+    return render_template('admin/update_inv.html', client_id=client_id, inv_to_update=inv_to_update, unpaid_list=unpaid_list, client_list=client_list, total_clients=total_clients, paid_inv=paid_inv, unpaid_inv=unpaid_inv, total_inv=total_inv,
                            billed_total=billed_total, collected_total=collected_total)
+
+
+
 
 
 @app.route('/dashboard/delete/<int:inv_id>', methods=['GET', 'POST'])
 def delete_invoice(inv_id: int):
     inv = InvoiceModel.query.get_or_404(inv_id)
-    try:
-        db.session.delete(inv)
-        db.session.commit()
-        flash('Invoice deleted successfully.')
-        return redirect(url_for('unpaid_view'))
-    except:
-        flash('Failed to delete invoice!')
+    if request == 'POST':
+        try:
+            db.session.delete(inv)
+            db.session.commit()
+            flash('Invoice deleted successfully.')
+            return redirect(url_for('unpaid_view'))
+        except:
+            flash('Failed to delete invoice!')
+            return redirect(url_for('unpaid_view'))
+
+
+
 
 
 @app.route('/make-inv', methods=['GET', 'POST'])
@@ -90,10 +129,16 @@ def make_inv():
                            total_clients=total_clients, billed_total=billed_total, collected_total=collected_total)
 
 
+
+
+
 @app.route('/add-client', methods=['GET', 'POST'])
 def add_client():
     return render_template('admin/add-client.html', paid_inv=paid_inv, unpaid_inv=unpaid_inv, total_inv=total_inv,
                            total_clients=total_clients, billed_total=billed_total, collected_total=collected_total)
+
+
+
 
 
 @app.route('/plot.png')
@@ -102,7 +147,6 @@ def plot_png():
     output = io.BytesIO()
     FigureCanvas(fig).print_png(output)
     return Response(output.getvalue(), mimetype='image/png')
-
 
 def create_figure():
     fig = Figure()
